requirements.txt
fastapi
uvicorn
python-multipart
itsdangerous
werkzeug
oracledb
requests
python-dotenv
zeep
Pillow
moviepy==1.0.3
PyMuPDF
numpy<2.0.0
chromadb==0.4.24
marshmallow==3.20.1